separated the emr_serverless.py module and the script run_emr_from_cli
