+++ eks_ml_pipeline/emr_serverless.py
@@ -1,4 +1,10 @@
-"""PySpark job on EMR Serverless"""
+"""
+PySpark job on EMR Serverless
+
+When run as a script 
+
+
+"""
 
 import gzip
 import argparse
@@ -179,11 +185,10 @@ class EMRServerless:
         if job_run_id is None:
             response = self.client.cancel_job_run(
                 applicationId=self.application_id, jobRunId=self.job_run_id
-            )
+                )
         else:
             response = self.client.cancel_job_run(
-                applicationId=self.application_id, jobRunId=job_run_id
-            )
+                applicationId=self.application_id, jobRunId=job_run_id)
         print('Successfully canceled job')
         return response.get("jobRun")
 
@@ -206,82 +211,3 @@ class EMRServerless:
             file_content = ""
         return print("File output from stdout.gz:\n----\n", str(file_content),
                      "\n----")
-
-
-def parse_args():
-    parser = argparse.ArgumentParser()
-    required_named = parser.add_argument_group(
-        "required named arguments"
-        )  # to display hyphen-prefixed args under "required arguments" group
-    required_named.add_argument(
-        "--job-role-arn", help="EMR Serverless IAM Job Role ARN", required=False
-        )
-    required_named.add_argument(
-        "--applicationId",
-        help="ApplicationId",
-        required=True,
-        )
-    required_named.add_argument(
-        "--s3-bucket",
-        help="S3 Bucket to use for logs and job output",
-        required=True,
-        )
-    required_named.add_argument(
-        "--entry-point",
-        help="Entry point to EMR serverless",
-        required=True,
-        )
-    required_named.add_argument(
-        "--zipped-env",
-        help = ("Path to the custom spark and python environemnt to use, "
-                + "with all the dependencies installed"),
-        required=True,
-        )
-    required_named.add_argument(
-        "--custom-spark-config",
-        help="Custom spark config",
-        required=False,
-        )
-
-    return parser.parse_args()
-
-
-if __name__ == "__main__":
-    # Require s3 bucket and job role to be provided
-    args = parse_args()
-    serverless_job_role_arn = args.job_role_arn
-    ap_id = args.applicationId
-    s3_bucket = args.s3_bucket
-    emr_emtry_point = args.entry_point
-    zipped_env_path = args.zipped_env
-    custom_spark_config = args.custom_spark_config
-
-    # Create and start a new EMRServerless Spark Application
-    emr_serverless = EMRServerless()
-
-    print("Starting EMR Serverless Spark App")
-#     uncomment below lines of code if you want to create a new application
-#     emr_serverless.create_application(
-#         "pd-autoencoder-test-emr-cli",
-#         "emr-6.6.0")
-    emr_serverless.start_application(ap_id)
-    print(emr_serverless)
-
-    # Run (and wait for) a Spark job
-    print("Submitting new Spark job")
-    job_id = emr_serverless.run_spark_job(
-        script_location = emr_emtry_point,
-        job_role_arn = serverless_job_role_arn,
-        application_id = ap_id,
-        arguments = [f"s3://{s3_bucket}/emr_serverless/output"],
-        s3_bucket_name = s3_bucket,
-        zipped_env_path = zipped_env_path,
-        custom_spark_config = custom_spark_config,
-    )
-    job_status = emr_serverless.get_job_run()
-    print(f"Job finished: {job_id}, status is: {job_status.get('state')}")
-
-    # Fetch and print the logs
-    spark_driver_logs = emr_serverless.fetch_driver_log(s3_bucket)
-
-    print("Done!")
